refactor: move level comments to new structure
